Added my version of the project
The main file is a User Interface, currently only the exit feature works along with a basic WIP for the search feature. I implemented a search function that can properly search by using EMPID or Name, I still need to add the search by SSN feature but we still have to add that to our columns. Also the output is still a WIP, I want it to show different information. If there's an issue with it working make sure to add the mysql-connector-j-8.1.0.jar as a module to your project, also make sure to change the password variable in the SearchEmployee.java file to your password instead.